// gojs/go.d.ts
    export type BrushLike = Brush | string;
// ...
    export class TextBlock extends GraphObject {
        /**
         * A newly constructed TextBlock has no string to show; if it did, it would draw the text, wrapping if needed, in the default font using a black stroke.
         */
        constructor();

        /**Gets or sets the an array of possible options for a custom TextEditingTool.*/
        choices: Array<string>;

        /**Gets or sets whether or not this TextBlock allows in-place editing of the .text string by the user with the help of the TextEditingTool.*/
        editable: boolean;

        /**Gets or sets the function to call if a text edit made with the TextEditingTool is invalid.*/
        errorFunction: (tool: TextEditingTool, oldstr: string, newstr: string) => void;

        /**Gets or sets the current font settings.*/
        font: string;

        /**Gets or sets the fractional distance along the main shape of a Graduated Panel at which this kind of tick text should end.*/
        graduatedEnd: number;

        /**Gets or sets the function to convert from a value along a Graduated Panel to a string.*/
        graduatedFunction: (val: number) => string;

        /**Gets or sets the fractional distance along the main shape of a Graduated Panel at which this text should start.*/
        graduatedStart: number;

        /**Gets or sets how frequently this text should be drawn within a Graduated Panel, in multiples of the Panel.graduatedTickUnit.*/
        interval: number;

        /**Gets or sets whether or not the text allows or displays multiple lines or embedded newlines.*/
        isMultiline: boolean;

        /**Gets or sets whether or not the text has a strikethrough line (line-through).*/
        isStrikethrough: boolean;

        /**Gets or sets whether or not the text is underlined.*/
        isUnderline: boolean;

        /**This read-only property returns the total number of lines in this TextBlock, including lines created from returns and wrapping.*/
        lineCount: number;

        /**Gets or sets the maximum number of lines that this TextBlock may display.*/
        maxLines: number;

        /**This read-only property returns the natural bounds of this TextBlock in local coordinates, as determined by its .font and .text string.*/
        naturalBounds: Rect;

        /**Gets or sets how text that is too long to display should be handled.*/
        overflow: EnumValue;

        /**Gets or sets the Brush (or CSS color string) that describes the stroke (color) of the .font.*/
        stroke: BrushLike;

        /**Gets or sets the current text string.*/
        text: string;

        /**Gets or sets the current text alignment property.*/
        textAlign: string;

        /**Gets or sets the function that is called after the TextBlock's text has been edited by the TextEditingTool.*/
        textEdited: (tb: TextBlock, oldstr: string, newstr: string) => void;

        /**Gets or sets the HTML Element that this TextBlock uses as its text editor in the TextEditingTool.*/
        textEditor: HTMLInfo | Element;  // Element is deprecated

        /**Gets or sets the predicate that determines whether or not a string of text is valid.*/
        textValidation: (tb: TextBlock, oldstr: string, newstr: string) => boolean;

        /**Gets or sets the vertical alignment when the TextBlock has more available vertical space than it needs to draw.*/
        verticalAlignment: Spot;

        /**Gets or sets whether the text should be wrapped if it is too long to fit on one line.*/
        wrap: EnumValue;

        /**The TextBlock will not wrap its text.*/
        static None: EnumValue;

        /** Used as the default value for TextBlock.overflow: if the width is too small to display all text, the TextBlock will clip.*/
        static OverflowClip: EnumValue;

        /** Used as a value for TextBlock.overflow: if the width is too small to display all text, the TextBlock will display an ellipsis.*/
        static OverflowEllipsis: EnumValue;

        /**The TextBlock will wrap text and the width of the TextBlock will be the desiredSize's width, if any.*/
        static WrapDesiredSize: EnumValue;

        /**The TextBlock will wrap text, making the width of the TextBlock equal to the width of the longest line.*/
        static WrapFit: EnumValue;

        spacingAbove: number;  // undocumented
        spacingBelow: number;  // undocumented
        flip: EnumValue;  // undocumented
        static getEllipsis(): string;  // undocumented
        static setEllipsis(val: string): void;  // undocumented
        static isValidFont(font: string): boolean;  // undocumented
    }
// ...
    export class Size {
        /**
         * The default constructor produces the Size(0,0).
         */
        constructor();

        /**
         * The two-argument constructor produces the Size(w,h).
         * @param {number} w this number must not be negative.
         * @param {number} h this number must not be negative.
         */
        constructor(w: number, h: number);

        /**Gets or sets the height value of the Size.*/
        height: number;

        /**Gets or sets the width value of the Size.*/
        width: number;

        /**
         * Create a copy of this Size, with the same values.
         */
        copy(): Size;

        /**
         * Indicates whether the given Size is equal to the current Size.
         * @param {Size} s The Size to compare to the current Size.
         * false otherwise.
         */
        equals(s: Size): boolean;

        /**
         * Indicates whether the given size is equal to this Size.
         * @param {number} w the width.
         * @param {number} h the height.
         * false otherwise.
         */
        equalTo(w: number, h: number): boolean;

        /**
         * True if this Size has Width and Height values that are real numbers and not infinity.
         */
        isReal(): boolean;

        /**
         * This static function can be used to read in a Size from a string that was produced by Size.stringify.
         * @param {string} str
         */
        static parse(str: string): Size;

        /**
         * Modify this Size so that its Width and Height values are the same as the given Size.
         * @param {Size} s the given Size.
         */
        set(s: Size): Size;

        /**
         * Modify this Size with new Width and Height values.
         * @param {number} w the width.
         * @param {number} h the height.
         */
        setTo(w: number, h: number): Size;

        /**
         * This static function can be used to write out a Size as a string that can be read by Size.parse.
         * @param {Size} val
         */
        static stringify(val: Size): string;

        equalsApprox(s: Size): boolean;  // undocumented
    }
// ...
    export class HTMLInfo {
        /**
         * Newly constructed HTMLInfo can be assigned as a value of TextEditingTool.defaultTextEditor, TextBlock.textEditor, GraphObject.contextMenu, Diagram.contextMenu, GraphObject.toolTip, or Diagram.toolTip.
         */
        constructor();

        /**Gets or sets the function to call when an HTMLInfo is shown.*/
        show: (obj: GraphObject, diagram: Diagram, arg: Tool) => void;
        /**Gets or sets the function to call when an HTMLInfo is to be hidden.*/
        hide: (diagram: Diagram, arg: Tool) => void;
        /**Gets or sets the primary HTML Element that represents this HTMLInfo.*/
        mainElement: Element;
        /**Gets or sets whether this HTMLInfo is modal. If true, clicking on the diagram after activation will automatically stop the tool associated with this HTMLInfo.*/
        isModal: boolean;
        /**Gets or sets a function that returns the primary value associated with this HTMLInfo, such as the value of a text editor, which would be solicited by the TextEditingTool.*/
        valueFunction: () => any;
    }
// ...
    export class TextEditingTool extends Tool {
        /**
         * You do not normally need to create an instance of this tool because one already exists as the ToolManager.textEditingTool, which you can modify.
         */
        constructor();

        /**Gets or sets the HTMLInfo or HTML Element that edits the text.*/
        currentTextEditor: HTMLInfo | Element;  // Element is deprecated

        /**Gets or sets the HTMLInfo or HTML Element that edits the text.*/
        defaultTextEditor: HTMLInfo | Element;  // Element is deprecated

        /**Gets or sets how user gestures can start in-place editing of text.*/
        starting: EnumValue;

        /**Gets or sets the TextBlock that is being edited.*/
        textBlock: TextBlock;

        /**Gets or sets the predicate that determines whether or not a string of text is valid.*/
        textValidation: (tb: TextBlock, oldstr: string, newstr: string) => boolean;

        /**
         * Finish editing by trying to accept the new text.
         * @param {EnumValue} reason The reason must be either TextEditingTool.LostFocus,
         * TextEditingTool.MouseDown, TextEditingTool.Tab, or TextEditingTool.Enter.
         */
        acceptText(reason: EnumValue): void;

        /**
         * This may run when there is a mouse-click on a TextBlock for which the TextBlock.editable property is true in a Part that Part.isSelected.
         */
        canStart(): boolean;

        /**
         * Start editing the text for a TextBlock.
         */
        doActivate(): void;

        /**
         * Abort any text editing operation.
         */
        doCancel(): void;

        /**
         * Release the mouse.
         */
        doDeactivate(): void;

        /**
         * A click (mouse up) calls TextEditingTool.doActivate if this tool is not already active and if TextEditingTool.canStart returns true.
         */
        doMouseDown(): void;

        /**
         * A click (mouse up) calls TextEditingTool.doActivate if this tool is not already active and if TextEditingTool.canStart returns true.
         */
        doMouseUp(): void;

        /**
         * This calls TextEditingTool.doActivate if there is a TextBlock supplied.
         */
        doStart(): void;

        /**
         * This predicate checks any TextBlock.textValidation predicate and this tool's .textValidation predicate to make sure the TextBlock.text property may be set to the new string.
         * @param {TextBlock} textblock the TextBlock that is being edited.
         * @param {string} oldstr the previous string value.
         * @param {string} newstr the proposed new string value.
         */
        isValidText(textblock: TextBlock, oldstr: string, newstr: string): boolean;

        /**The user has typed ENTER.*/
        static Enter: EnumValue;

        /**The text editing control has lost focus.*/
        static LostFocus: EnumValue;

        /**The user has clicked somewhere else in the diagram.*/
        static MouseDown: EnumValue;

        /**A double click on a TextBlock with TextBlock.editable property set to true will start in-place editing.*/
        static DoubleClick: EnumValue;

        /**A single click on a TextBlock with TextBlock.editable property set to true will start in-place editing.*/
        static SingleClick: EnumValue;

        /**A single click on a TextBlock with TextBlock.editable property set to true will start in-place editing, but only if the Part that the TextBlock is in is already selected.*/
        static SingleClickSelected: EnumValue;

        /**The user has typed TAB.*/
        static Tab: EnumValue;
    }
// ...
    export class EnumValue {  // undocumented
        //Rawr!
    }